requirements.txt
pandas
numpy
scikit-learn
matplotlib
sqlalchemy
psycopg2
python-dotenv
joblib